requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
streamlit==1.39.0
torch==2.4.0
torch-geometric==2.6.1
networkx==3.4.2
pandas==2.2.3
scikit-learn==1.5.2
mlflow==2.17.0
dvc==3.55.0
pyyaml==6.0.2
python-dotenv==1.0.1
rdkit==2024.9.5
evidently==0.4.40